Adding Vue Resources and Initial Pages.
