+++ BackEnd/.gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Mon Mar 17 12:14:03 CDT 2025
+gradle.version=8.12.1
