+++ FrontEnd/frontend/src/components/Sidebar.vue
@@ -1,0 +1,20 @@
+<script>
+  export default {
+
+  }
+</script>
+
+<template>
+  
+</template>
+
+<style scoped>
+  .sidebar {
+    margin: 0;
+    width: 350px;
+    height: 100%;
+    text-align: left;
+    background: #2c3e50;
+    display: inline-block;
+  }
+</style>
